network availability changes
Added checks in the IOManager so that it checks if a network is available before executing commands to elastic search. Also added methods to the manager so that all IO can now go right through the IOManager which acts as a proxy for both the LocalDataManager and the NetworkDataManager. Fixed annoying mistake where extra activities were spawned before exiting a current activity. Added a method to clear the users local data on logout, but left it out for now until pulling of data from network is done for all activities