uploaded elastic search index creation file
uploaded the elastic search index creation json body, so that in the event that some asshole deletes our index we can quickly recreate it on the fly. Also helps in case we want to change the properties of fields in or mappings since elastic search doesn't allow this, and thus you have to delete the index and recreate it.