major changes to iomananger, network, and elastic search
added much functionality and refactors for using the network manager along with the elastic search controller to make it all more consistent and add methods for other functionality needed by the application. Moved the elastic search controller test to the androidTest folder so that it can access the androidSDK during test run. Deleted habit location because it was not needed anymore, moved the lat long coord field of this class to the habitevent class.